fix crasher
